refactor(state): split store implementation by concern
Move the DB store records, migration setup, request handling, catalog logic, identity/audit helpers, and conversion utilities into focused files while keeping the state package API unchanged.
