+++ internal/state/store.go
@@ -1,24 +1,8 @@
 package state
 
 import (
-	"bytes"
-	"encoding/json"
 	"errors"
-	"fmt"
-	"log/slog"
-	"os"
-	"path/filepath"
-	"sort"
-	"strings"
-	"sync"
 	"time"
-
-	"github.com/glebarez/sqlite"
-	"github.com/qiniu/ci-runner/internal/labelutil"
-	"gorm.io/driver/postgres"
-	"gorm.io/gorm"
-	"gorm.io/gorm/clause"
-	gormlogger "gorm.io/gorm/logger"
 )
 
 const (
@@ -220,1776 +204,4 @@ type Options struct {
 	Backend        string
 	DatabaseURL    string
 	MigrateOnStart bool
-}
-
-type DBStore struct {
-	opts     Options
-	mu       sync.Mutex
-	db       *gorm.DB
-	migrated bool
-}
-
-type runnerRequestRecord struct {
-	ID                  string     `gorm:"column:id;primaryKey"`
-	Source              string     `gorm:"column:source;not null"`
-	WorkflowJobID       *int64     `gorm:"column:workflow_job_id;uniqueIndex:idx_runner_requests_workflow_job_id"`
-	RepositoryFullName  string     `gorm:"column:repository_full_name"`
-	RequestedLabelsJSON string     `gorm:"column:requested_labels_json"`
-	LabelsJSON          string     `gorm:"column:labels_json;not null"`
-	ProfileName         string     `gorm:"column:profile_name"`
-	RunnerGroup         string     `gorm:"column:runner_group"`
-	RunnerName          string     `gorm:"column:runner_name;not null"`
-	Status              string     `gorm:"column:status;not null;index:idx_runner_requests_status_updated;index:idx_runner_requests_status_retry_queue;index:idx_runner_requests_lease_expiry"`
-	FailureStage        string     `gorm:"column:failure_stage"`
-	FailureReason       string     `gorm:"column:failure_reason"`
-	LastErrorCode       string     `gorm:"column:last_error_code"`
-	LastErrorMessage    string     `gorm:"column:last_error_message"`
-	LastErrorRetryable  bool       `gorm:"column:last_error_retryable;not null;default:false"`
-	RetryCount          int        `gorm:"column:retry_count;not null;default:0"`
-	SandboxID           string     `gorm:"column:sandbox_id"`
-	ProcessPID          uint32     `gorm:"column:process_pid"`
-	AssignedJobID       int64      `gorm:"column:assigned_job_id"`
-	AssignedJobName     string     `gorm:"column:assigned_job_name"`
-	Error               string     `gorm:"column:error"`
-	GitHubPayloadJSON   string     `gorm:"column:github_payload_json"`
-	QueuedAt            time.Time  `gorm:"column:queued_at;not null;index:idx_runner_requests_status_updated;index:idx_runner_requests_status_retry_queue"`
-	LastAttemptAt       *time.Time `gorm:"column:last_attempt_at"`
-	NextRetryAt         *time.Time `gorm:"column:next_retry_at;index:idx_runner_requests_status_retry_queue"`
-	CreatingAt          *time.Time `gorm:"column:creating_at"`
-	RunningAt           *time.Time `gorm:"column:running_at"`
-	StoppingAt          *time.Time `gorm:"column:stopping_at"`
-	CompletedAt         *time.Time `gorm:"column:completed_at"`
-	FailedAt            *time.Time `gorm:"column:failed_at"`
-	LeaseOwner          string     `gorm:"column:lease_owner"`
-	LeaseExpiresAt      *time.Time `gorm:"column:lease_expires_at;index:idx_runner_requests_lease_expiry"`
-	UpdatedAt           time.Time  `gorm:"column:updated_at;not null;index:idx_runner_requests_status_updated"`
-	Version             int64      `gorm:"column:version;not null;default:0"`
-}
-
-func (runnerRequestRecord) TableName() string { return "runner_requests" }
-
-type runnerEventRecord struct {
-	ID          int64     `gorm:"column:id;primaryKey;autoIncrement"`
-	RequestID   string    `gorm:"column:request_id;not null;index:idx_runner_events_request_created"`
-	EventType   string    `gorm:"column:event_type;not null"`
-	Stage       string    `gorm:"column:stage"`
-	Message     string    `gorm:"column:message"`
-	PayloadJSON string    `gorm:"column:payload_json"`
-	CreatedAt   time.Time `gorm:"column:created_at;not null;index:idx_runner_events_request_created"`
-}
-
-func (runnerEventRecord) TableName() string { return "runner_events" }
-
-type runnerProfileRecord struct {
-	Name             string    `gorm:"column:name;primaryKey"`
-	LabelsJSON       string    `gorm:"column:labels_json;not null"`
-	TemplateID       string    `gorm:"column:template_id;not null"`
-	RunnerGroup      string    `gorm:"column:runner_group"`
-	MaxConcurrency   int       `gorm:"column:max_concurrency;not null"`
-	MinIdle          int       `gorm:"column:min_idle;not null;default:0"`
-	Priority         int       `gorm:"column:priority;not null;default:0"`
-	Enabled          bool      `gorm:"column:enabled;not null"`
-	DefaultAvailable bool      `gorm:"column:default_available;not null"`
-	CreatedAt        time.Time `gorm:"column:created_at;not null"`
-	UpdatedAt        time.Time `gorm:"column:updated_at;not null"`
-}
-
-func (runnerProfileRecord) TableName() string { return "runner_profiles" }
-
-type legacyRunnerProfileDefaultAvailableColumn struct {
-	DefaultAvailable bool `gorm:"column:default_available;not null;default:true"`
-}
-
-func (legacyRunnerProfileDefaultAvailableColumn) TableName() string { return "runner_profiles" }
-
-type runnerGroupRecord struct {
-	Name        string    `gorm:"column:name;primaryKey"`
-	Description string    `gorm:"column:description"`
-	Enabled     bool      `gorm:"column:enabled;not null"`
-	CreatedAt   time.Time `gorm:"column:created_at;not null"`
-	UpdatedAt   time.Time `gorm:"column:updated_at;not null"`
-}
-
-func (runnerGroupRecord) TableName() string { return "runner_groups" }
-
-type runnerGroupSpecRecord struct {
-	GroupName string    `gorm:"column:group_name;primaryKey"`
-	SpecName  string    `gorm:"column:spec_name;primaryKey;index:idx_runner_group_specs_spec"`
-	CreatedAt time.Time `gorm:"column:created_at;not null"`
-}
-
-func (runnerGroupSpecRecord) TableName() string { return "runner_group_specs" }
-
-type repositoryPolicyRecord struct {
-	ID                 int64     `gorm:"column:id;primaryKey;autoIncrement"`
-	RepositoryFullName string    `gorm:"column:repository_full_name;not null;uniqueIndex:idx_repository_policies_repository_profile,where:profile_name <> '';uniqueIndex:idx_repository_policies_repository_group,where:runner_group_name <> ''"`
-	ProfileName        string    `gorm:"column:profile_name;not null;uniqueIndex:idx_repository_policies_repository_profile,where:profile_name <> ''"`
-	RunnerGroupName    string    `gorm:"column:runner_group_name;not null;default:'';uniqueIndex:idx_repository_policies_repository_group,where:runner_group_name <> ''"`
-	Enabled            bool      `gorm:"column:enabled;not null"`
-	CreatedAt          time.Time `gorm:"column:created_at;not null"`
-}
-
-func (repositoryPolicyRecord) TableName() string { return "repository_policies" }
-
-type legacyRepositoryPolicyRunnerGroupNameColumn struct {
-	RunnerGroupName string `gorm:"column:runner_group_name;not null;default:''"`
-}
-
-func (legacyRepositoryPolicyRunnerGroupNameColumn) TableName() string { return "repository_policies" }
-
-type auditEventRecord struct {
-	ID           int64     `gorm:"column:id;primaryKey;autoIncrement"`
-	Actor        string    `gorm:"column:actor;not null"`
-	Action       string    `gorm:"column:action;not null"`
-	ResourceType string    `gorm:"column:resource_type;not null"`
-	ResourceID   string    `gorm:"column:resource_id;not null"`
-	PayloadJSON  string    `gorm:"column:payload_json"`
-	CreatedAt    time.Time `gorm:"column:created_at;not null;index:idx_audit_events_created"`
-}
-
-func (auditEventRecord) TableName() string { return "audit_events" }
-
-type accountRecord struct {
-	ID        int64     `gorm:"column:id;primaryKey;autoIncrement"`
-	Role      string    `gorm:"column:role;not null"`
-	CreatedAt time.Time `gorm:"column:created_at;not null"`
-	UpdatedAt time.Time `gorm:"column:updated_at;not null"`
-}
-
-func (accountRecord) TableName() string { return "accounts" }
-
-type oauthIdentityRecord struct {
-	ID            int64         `gorm:"column:id;primaryKey;autoIncrement"`
-	AccountID     int64         `gorm:"column:account_id;not null;index:idx_oauth_identities_account"`
-	Account       accountRecord `gorm:"foreignKey:AccountID;constraint:OnDelete:CASCADE"`
-	OAuthProvider string        `gorm:"column:oauth_provider;not null;uniqueIndex:idx_oauth_identities_provider_subject"`
-	OAuthSubject  string        `gorm:"column:oauth_subject;not null;uniqueIndex:idx_oauth_identities_provider_subject"`
-	OAuthLogin    string        `gorm:"column:oauth_login;not null"`
-	CreatedAt     time.Time     `gorm:"column:created_at;not null"`
-	UpdatedAt     time.Time     `gorm:"column:updated_at;not null"`
-}
-
-func (oauthIdentityRecord) TableName() string { return "oauth_identities" }
-
-func New(dir string) Store {
-	return NewWithOptions(Options{
-		Backend:        BackendSQLite,
-		DatabaseURL:    filepath.Join(dir, "runnerd.db"),
-		MigrateOnStart: true,
-	})
-}
-
-func NewWithOptions(opts Options) Store {
-	backend := strings.ToLower(strings.TrimSpace(opts.Backend))
-	if backend == "" {
-		backend = BackendSQLite
-	}
-	if opts.DatabaseURL == "" && backend == BackendSQLite {
-		opts.DatabaseURL = filepath.Join(".", "var", "runnerd.db")
-	}
-	opts.Backend = backend
-	return &DBStore{opts: opts}
-}
-
-func (s *DBStore) Ensure() error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.db != nil {
-		return nil
-	}
-
-	db, err := s.open()
-	if err != nil {
-		return err
-	}
-	if s.opts.MigrateOnStart && !s.migrated {
-		if err := s.migrate(db); err != nil {
-			return err
-		}
-		s.migrated = true
-	}
-	s.db = db
-	return nil
-}
-
-func (s *DBStore) CreateRequest(req RunnerRequest, payload []byte) (bool, RunnerState, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return false, RunnerState{}, err
-	}
-
-	req.ID = sanitizeID(req.ID)
-	if req.ID == "" {
-		return false, RunnerState{}, fmt.Errorf("request id is empty")
-	}
-	req.RunnerName = sanitizeRunnerName(req.RunnerName)
-	if req.RunnerName == "" {
-		req.RunnerName = "e2b-" + req.ID
-	}
-	now := time.Now().UTC()
-	if req.CreatedAt.IsZero() {
-		req.CreatedAt = now
-	}
-
-	if len(req.RequestedLabels) == 0 {
-		req.RequestedLabels = append([]string(nil), req.Labels...)
-	}
-	labelsJSON, err := json.Marshal(req.Labels)
-	if err != nil {
-		return false, RunnerState{}, err
-	}
-	requestedLabelsJSON, err := json.Marshal(req.RequestedLabels)
-	if err != nil {
-		return false, RunnerState{}, err
-	}
-	record := runnerRequestRecord{
-		ID:                  req.ID,
-		Source:              req.Source,
-		RepositoryFullName:  req.RepositoryFullName,
-		RequestedLabelsJSON: string(requestedLabelsJSON),
-		LabelsJSON:          string(labelsJSON),
-		ProfileName:         req.ProfileName,
-		RunnerGroup:         req.RunnerGroup,
-		RunnerName:          req.RunnerName,
-		Status:              StatusQueued,
-		GitHubPayloadJSON:   string(payload),
-		QueuedAt:            req.CreatedAt,
-		UpdatedAt:           now,
-		Version:             0,
-	}
-	if req.JobID != 0 {
-		record.WorkflowJobID = &req.JobID
-	}
-	result := db.Clauses(clause.OnConflict{DoNothing: true}).Create(&record)
-	if result.Error != nil {
-		return false, RunnerState{}, result.Error
-	}
-	if result.RowsAffected == 0 {
-		st, err := s.ReadState(req.ID)
-		return false, st, err
-	}
-	return true, recordToState(record), nil
-}
-
-func (s *DBStore) ReadRequest(id string) (RunnerRequest, error) {
-	record, err := s.readRecord(id)
-	if err != nil {
-		return RunnerRequest{}, err
-	}
-	return recordToRequest(record)
-}
-
-func (s *DBStore) ReadState(id string) (RunnerState, error) {
-	record, err := s.readRecord(id)
-	if err != nil {
-		return RunnerState{}, err
-	}
-	return recordToState(record), nil
-}
-
-func (s *DBStore) WriteState(st RunnerState) error {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return err
-	}
-	st.ID = sanitizeID(st.ID)
-	if st.ID == "" {
-		return fmt.Errorf("request id is empty")
-	}
-	now := time.Now().UTC()
-	st.UpdatedAt = now
-	applyStateTimestamps(&st, now)
-
-	updates := map[string]any{
-		"status":               st.Status,
-		"repository_full_name": st.RepositoryFullName,
-		"profile_name":         st.ProfileName,
-		"runner_group":         st.RunnerGroup,
-		"runner_name":          sanitizeRunnerName(st.RunnerName),
-		"sandbox_id":           st.SandboxID,
-		"process_pid":          st.ProcessPID,
-		"assigned_job_id":      st.AssignedJobID,
-		"assigned_job_name":    st.AssignedJobName,
-		"error":                st.Error,
-		"failure_stage":        st.FailureStage,
-		"failure_reason":       st.FailureReason,
-		"last_error_code":      st.LastErrorCode,
-		"last_error_message":   st.LastErrorMessage,
-		"last_error_retryable": st.LastErrorRetryable,
-		"retry_count":          st.RetryCount,
-		"updated_at":           st.UpdatedAt,
-		"version":              st.Version + 1,
-		"last_attempt_at":      zeroTimeToPointer(st.LastAttemptAt),
-		"next_retry_at":        zeroTimeToPointer(st.NextRetryAt),
-		"creating_at":          zeroTimeToPointer(st.CreatingAt),
-		"running_at":           zeroTimeToPointer(st.RunningAt),
-		"stopping_at":          zeroTimeToPointer(st.StoppingAt),
-		"completed_at":         zeroTimeToPointer(st.CompletedAt),
-		"failed_at":            zeroTimeToPointer(st.FailedAt),
-		"lease_owner":          st.LeaseOwner,
-		"lease_expires_at":     zeroTimeToPointer(st.LeaseExpiresAt),
-	}
-	result := db.Model(&runnerRequestRecord{}).
-		Where("id = ? AND version = ?", st.ID, st.Version).
-		Updates(updates)
-	if result.Error != nil {
-		return result.Error
-	}
-	if result.RowsAffected == 0 {
-		return ErrConflict
-	}
-	return nil
-}
-
-func (s *DBStore) ListStates() ([]RunnerState, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return nil, err
-	}
-	var records []runnerRequestRecord
-	if err := db.Order("queued_at DESC").Find(&records).Error; err != nil {
-		return nil, err
-	}
-	states := make([]RunnerState, 0, len(records))
-	for _, record := range records {
-		states = append(states, recordToState(record))
-	}
-	return states, nil
-}
-
-func (s *DBStore) ListActiveStates() ([]RunnerState, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return nil, err
-	}
-	var records []runnerRequestRecord
-	if err := db.
-		Where("status IN ?", activeStatuses()).
-		Order("queued_at DESC").
-		Find(&records).Error; err != nil {
-		return nil, err
-	}
-	states := make([]RunnerState, 0, len(records))
-	for _, record := range records {
-		states = append(states, recordToState(record))
-	}
-	return states, nil
-}
-
-func (s *DBStore) ListMismatchedCompletedStates(limit int) ([]RunnerState, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return nil, err
-	}
-	if limit <= 0 {
-		limit = 50
-	}
-	var records []runnerRequestRecord
-	if err := db.
-		Where("status = ?", StatusCompleted).
-		Where("workflow_job_id IS NOT NULL AND workflow_job_id != 0").
-		Where("assigned_job_id != 0 AND assigned_job_id != workflow_job_id").
-		Order("updated_at DESC").
-		Limit(limit).
-		Find(&records).Error; err != nil {
-		return nil, err
-	}
-	states := make([]RunnerState, 0, len(records))
-	for _, record := range records {
-		states = append(states, recordToState(record))
-	}
-	return states, nil
-}
-
-func (s *DBStore) ListFailedWorkflowJobStates(limit int) ([]RunnerState, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return nil, err
-	}
-	if limit <= 0 {
-		limit = 50
-	}
-	var records []runnerRequestRecord
-	if err := db.
-		Where("status = ?", StatusFailed).
-		Where("workflow_job_id IS NOT NULL AND workflow_job_id != 0").
-		Where("failure_stage IN ?", []string{"recovery", "cleanup"}).
-		Order("updated_at DESC").
-		Limit(limit).
-		Find(&records).Error; err != nil {
-		return nil, err
-	}
-	states := make([]RunnerState, 0, len(records))
-	for _, record := range records {
-		states = append(states, recordToState(record))
-	}
-	return states, nil
-}
-
-func (s *DBStore) ListStatesPage(limit, offset int) ([]RunnerState, int64, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return nil, 0, err
-	}
-	if limit <= 0 {
-		limit = 100
-	}
-	if offset < 0 {
-		offset = 0
-	}
-	var total int64
-	if err := db.Model(&runnerRequestRecord{}).Count(&total).Error; err != nil {
-		return nil, 0, err
-	}
-	var records []runnerRequestRecord
-	if err := db.Order("queued_at DESC").Limit(limit).Offset(offset).Find(&records).Error; err != nil {
-		return nil, 0, err
-	}
-	states := make([]RunnerState, 0, len(records))
-	for _, record := range records {
-		states = append(states, recordToState(record))
-	}
-	return states, total, nil
-}
-
-func activeStatuses() []string {
-	return []string{StatusQueued, StatusCreating, StatusRunning, StatusStopping}
-}
-
-func (s *DBStore) ActiveCount() (int, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return 0, err
-	}
-	var count int64
-	if err := db.Model(&runnerRequestRecord{}).
-		Where("status IN ?", activeStatuses()).
-		Count(&count).Error; err != nil {
-		return 0, err
-	}
-	return int(count), nil
-}
-
-func (s *DBStore) InFlightCount() (int, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return 0, err
-	}
-	var count int64
-	if err := db.Model(&runnerRequestRecord{}).
-		Where("status IN ?", []string{StatusCreating, StatusRunning, StatusStopping}).
-		Count(&count).Error; err != nil {
-		return 0, err
-	}
-	return int(count), nil
-}
-
-func (s *DBStore) ActiveCountForProfile(name string) (int, error) {
-	return s.countStatesForProfile(name, []string{StatusQueued, StatusCreating, StatusRunning, StatusStopping})
-}
-
-func (s *DBStore) InFlightCountForProfile(name string) (int, error) {
-	return s.countStatesForProfile(name, []string{StatusCreating, StatusRunning, StatusStopping})
-}
-
-func (s *DBStore) ClaimNextRunnable(workerID string, now time.Time, leaseTTL time.Duration) (RunnerRequest, RunnerState, bool, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return RunnerRequest{}, RunnerState{}, false, err
-	}
-	workerID = strings.TrimSpace(workerID)
-	if workerID == "" {
-		return RunnerRequest{}, RunnerState{}, false, fmt.Errorf("worker id is required")
-	}
-	leaseExpiresAt := now.UTC().Add(leaseTTL)
-	var claimed runnerRequestRecord
-	err = db.Transaction(func(tx *gorm.DB) error {
-		var candidate runnerRequestRecord
-		if err := tx.
-			Where("status = ?", StatusQueued).
-			Where("(next_retry_at IS NULL OR next_retry_at <= ?)", now.UTC()).
-			Where("(lease_expires_at IS NULL OR lease_expires_at <= ?)", now.UTC()).
-			Order("queued_at ASC").
-			First(&candidate).Error; err != nil {
-			return err
-		}
-		result := tx.Model(&runnerRequestRecord{}).
-			Where("id = ?", candidate.ID).
-			Where("status = ?", StatusQueued).
-			Where("(lease_expires_at IS NULL OR lease_expires_at <= ?)", now.UTC()).
-			Updates(map[string]any{
-				"lease_owner":      workerID,
-				"lease_expires_at": leaseExpiresAt,
-				"last_attempt_at":  now.UTC(),
-				"updated_at":       now.UTC(),
-				"version":          candidate.Version + 1,
-			})
-		if result.Error != nil {
-			return result.Error
-		}
-		if result.RowsAffected == 0 {
-			return ErrConflict
-		}
-		if err := tx.First(&claimed, "id = ?", candidate.ID).Error; err != nil {
-			return err
-		}
-		return nil
-	})
-	if errors.Is(err, gorm.ErrRecordNotFound) {
-		return RunnerRequest{}, RunnerState{}, false, nil
-	}
-	if errors.Is(err, ErrConflict) {
-		return RunnerRequest{}, RunnerState{}, false, nil
-	}
-	if err != nil {
-		return RunnerRequest{}, RunnerState{}, false, err
-	}
-	req, err := recordToRequest(claimed)
-	if err != nil {
-		return RunnerRequest{}, RunnerState{}, false, err
-	}
-	return req, recordToState(claimed), true, nil
-}
-
-func (s *DBStore) ReleaseLease(id, workerID string) error {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return err
-	}
-	return db.Model(&runnerRequestRecord{}).
-		Where("id = ? AND lease_owner = ?", sanitizeID(id), strings.TrimSpace(workerID)).
-		Updates(map[string]any{
-			"lease_owner":      "",
-			"lease_expires_at": nil,
-		}).Error
-}
-
-func (s *DBStore) RetryRequest(id string, now time.Time) (RunnerState, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return RunnerState{}, err
-	}
-	id = sanitizeID(id)
-	var saved runnerRequestRecord
-	if err := db.Transaction(func(tx *gorm.DB) error {
-		var record runnerRequestRecord
-		if err := tx.First(&record, "id = ?", id).Error; err != nil {
-			return err
-		}
-		switch record.Status {
-		case StatusFailed:
-		case StatusQueued:
-			if record.NextRetryAt == nil || !record.LastErrorRetryable {
-				return ErrRetryNotAllowed
-			}
-		default:
-			return ErrRetryNotAllowed
-		}
-		record.Status = StatusQueued
-		record.Error = ""
-		record.FailureStage = ""
-		record.FailureReason = ""
-		record.LastErrorCode = ""
-		record.LastErrorMessage = ""
-		record.LastErrorRetryable = false
-		record.SandboxID = ""
-		record.ProcessPID = 0
-		record.AssignedJobID = 0
-		record.AssignedJobName = ""
-		record.NextRetryAt = nil
-		record.LeaseOwner = ""
-		record.LeaseExpiresAt = nil
-		record.CreatingAt = nil
-		record.RunningAt = nil
-		record.StoppingAt = nil
-		record.CompletedAt = nil
-		record.FailedAt = nil
-		record.UpdatedAt = now.UTC()
-		record.Version++
-		if err := tx.Model(&runnerRequestRecord{}).
-			Where("id = ? AND version = ?", record.ID, record.Version-1).
-			Updates(map[string]any{
-				"status":               record.Status,
-				"error":                record.Error,
-				"failure_stage":        record.FailureStage,
-				"failure_reason":       record.FailureReason,
-				"last_error_code":      record.LastErrorCode,
-				"last_error_message":   record.LastErrorMessage,
-				"last_error_retryable": record.LastErrorRetryable,
-				"sandbox_id":           record.SandboxID,
-				"process_pid":          record.ProcessPID,
-				"assigned_job_id":      record.AssignedJobID,
-				"assigned_job_name":    record.AssignedJobName,
-				"next_retry_at":        nil,
-				"lease_owner":          record.LeaseOwner,
-				"lease_expires_at":     nil,
-				"creating_at":          nil,
-				"running_at":           nil,
-				"stopping_at":          nil,
-				"completed_at":         nil,
-				"failed_at":            nil,
-				"updated_at":           record.UpdatedAt,
-				"version":              record.Version,
-			}).Error; err != nil {
-			return err
-		}
-		if err := tx.First(&saved, "id = ?", id).Error; err != nil {
-			return err
-		}
-		return nil
-	}); err != nil {
-		return RunnerState{}, err
-	}
-	return recordToState(saved), nil
-}
-
-func (s *DBStore) AppendLog(id, name string, data []byte) {
-	if len(data) == 0 {
-		return
-	}
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		slog.Default().Warn("append runner log failed", "id", id, "name", name, "error", err)
-		return
-	}
-	eventType, err := logEventType(name)
-	if err != nil {
-		slog.Default().Warn("append runner log failed", "id", id, "name", name, "error", err)
-		return
-	}
-	if err := db.Create(&runnerEventRecord{
-		RequestID: id,
-		EventType: eventType,
-		Message:   string(data),
-		CreatedAt: time.Now().UTC(),
-	}).Error; err != nil {
-		slog.Default().Warn("append runner log failed", "id", id, "name", name, "error", err)
-	}
-}
-
-func (s *DBStore) ReadLog(id, name string, maxBytes int64) ([]byte, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return nil, err
-	}
-	eventType, err := logEventType(name)
-	if err != nil {
-		return nil, err
-	}
-	var events []runnerEventRecord
-	if err := db.Where("request_id = ? AND event_type = ?", sanitizeID(id), eventType).
-		Order("id ASC").
-		Find(&events).Error; err != nil {
-		return nil, err
-	}
-	if len(events) == 0 {
-		return nil, os.ErrNotExist
-	}
-	var buf bytes.Buffer
-	for _, event := range events {
-		buf.WriteString(event.Message)
-	}
-	data := buf.Bytes()
-	if maxBytes > 0 && int64(len(data)) > maxBytes {
-		data = data[len(data)-int(maxBytes):]
-	}
-	return append([]byte(nil), data...), nil
-}
-
-func (s *DBStore) ListProfiles() ([]RunnerProfile, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return nil, err
-	}
-	var records []runnerProfileRecord
-	if err := db.Order("priority DESC, name ASC").Find(&records).Error; err != nil {
-		return nil, err
-	}
-	profiles := make([]RunnerProfile, 0, len(records))
-	for _, record := range records {
-		profile, err := recordToProfile(record)
-		if err != nil {
-			return nil, err
-		}
-		profiles = append(profiles, profile)
-	}
-	return profiles, nil
-}
-
-func (s *DBStore) GetProfile(name string) (RunnerProfile, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return RunnerProfile{}, err
-	}
-	var record runnerProfileRecord
-	if err := db.First(&record, "name = ?", strings.TrimSpace(name)).Error; err != nil {
-		return RunnerProfile{}, err
-	}
-	return recordToProfile(record)
-}
-
-func (s *DBStore) UpsertProfile(profile RunnerProfile) (RunnerProfile, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return RunnerProfile{}, err
-	}
-	profile.Name = strings.TrimSpace(profile.Name)
-	if profile.Name == "" {
-		return RunnerProfile{}, fmt.Errorf("profile name is required")
-	}
-	labelsJSON, err := json.Marshal(profile.Labels)
-	if err != nil {
-		return RunnerProfile{}, err
-	}
-	now := time.Now().UTC()
-	record := runnerProfileRecord{
-		Name:             profile.Name,
-		LabelsJSON:       string(labelsJSON),
-		TemplateID:       profile.TemplateID,
-		RunnerGroup:      profile.RunnerGroup,
-		MaxConcurrency:   profile.MaxConcurrency,
-		MinIdle:          profile.MinIdle,
-		Priority:         profile.Priority,
-		Enabled:          profile.Enabled,
-		DefaultAvailable: profile.DefaultAvailable,
-		CreatedAt:        profile.CreatedAt,
-		UpdatedAt:        now,
-	}
-	if record.CreatedAt.IsZero() {
-		record.CreatedAt = now
-	}
-	if err := db.Clauses(clause.OnConflict{
-		Columns:   []clause.Column{{Name: "name"}},
-		DoUpdates: clause.Assignments(map[string]any{"labels_json": record.LabelsJSON, "template_id": record.TemplateID, "runner_group": record.RunnerGroup, "max_concurrency": record.MaxConcurrency, "min_idle": record.MinIdle, "priority": record.Priority, "enabled": record.Enabled, "default_available": record.DefaultAvailable, "updated_at": record.UpdatedAt}),
-	}).Create(&record).Error; err != nil {
-		return RunnerProfile{}, err
-	}
-	return s.GetProfile(record.Name)
-}
-
-func (s *DBStore) DeleteProfile(name string) error {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return err
-	}
-	name = strings.TrimSpace(name)
-	return db.Transaction(func(tx *gorm.DB) error {
-		if err := tx.Delete(&runnerGroupSpecRecord{}, "spec_name = ?", name).Error; err != nil {
-			return err
-		}
-		return tx.Delete(&runnerProfileRecord{}, "name = ?", name).Error
-	})
-}
-
-func (s *DBStore) ListRunnerGroups() ([]RunnerGroup, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return nil, err
-	}
-	var records []runnerGroupRecord
-	if err := db.Order("name ASC").Find(&records).Error; err != nil {
-		return nil, err
-	}
-	groups := make([]RunnerGroup, 0, len(records))
-	for _, record := range records {
-		group, err := s.recordToRunnerGroup(db, record)
-		if err != nil {
-			return nil, err
-		}
-		groups = append(groups, group)
-	}
-	return groups, nil
-}
-
-func (s *DBStore) GetRunnerGroup(name string) (RunnerGroup, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return RunnerGroup{}, err
-	}
-	var record runnerGroupRecord
-	if err := db.First(&record, "name = ?", strings.TrimSpace(name)).Error; err != nil {
-		return RunnerGroup{}, err
-	}
-	return s.recordToRunnerGroup(db, record)
-}
-
-func (s *DBStore) UpsertRunnerGroup(group RunnerGroup) (RunnerGroup, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return RunnerGroup{}, err
-	}
-	group.Name = strings.TrimSpace(group.Name)
-	if group.Name == "" {
-		return RunnerGroup{}, fmt.Errorf("runner group name is required")
-	}
-	specNames := uniqueTrimmed(group.SpecNames)
-	now := time.Now().UTC()
-	record := runnerGroupRecord{
-		Name:        group.Name,
-		Description: strings.TrimSpace(group.Description),
-		Enabled:     group.Enabled,
-		CreatedAt:   group.CreatedAt,
-		UpdatedAt:   now,
-	}
-	if record.CreatedAt.IsZero() {
-		record.CreatedAt = now
-	}
-	if err := db.Transaction(func(tx *gorm.DB) error {
-		for _, specName := range specNames {
-			var count int64
-			if err := tx.Model(&runnerProfileRecord{}).Where("name = ?", specName).Count(&count).Error; err != nil {
-				return err
-			}
-			if count == 0 {
-				return fmt.Errorf("runner spec %q does not exist", specName)
-			}
-		}
-		if err := tx.Clauses(clause.OnConflict{
-			Columns: []clause.Column{{Name: "name"}},
-			DoUpdates: clause.Assignments(map[string]any{
-				"description": record.Description,
-				"enabled":     record.Enabled,
-				"updated_at":  record.UpdatedAt,
-			}),
-		}).Create(&record).Error; err != nil {
-			return err
-		}
-		if err := tx.Delete(&runnerGroupSpecRecord{}, "group_name = ?", record.Name).Error; err != nil {
-			return err
-		}
-		for _, specName := range specNames {
-			link := runnerGroupSpecRecord{GroupName: record.Name, SpecName: specName, CreatedAt: now}
-			if err := tx.Clauses(clause.OnConflict{DoNothing: true}).Create(&link).Error; err != nil {
-				return err
-			}
-		}
-		return nil
-	}); err != nil {
-		return RunnerGroup{}, err
-	}
-	return s.GetRunnerGroup(record.Name)
-}
-
-func (s *DBStore) DeleteRunnerGroup(name string) error {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return err
-	}
-	name = strings.TrimSpace(name)
-	return db.Transaction(func(tx *gorm.DB) error {
-		if err := tx.Delete(&runnerGroupSpecRecord{}, "group_name = ?", name).Error; err != nil {
-			return err
-		}
-		if err := tx.Model(&repositoryPolicyRecord{}).
-			Where("runner_group_name = ?", name).
-			Update("enabled", false).Error; err != nil {
-			return err
-		}
-		return tx.Delete(&runnerGroupRecord{}, "name = ?", name).Error
-	})
-}
-
-func (s *DBStore) ListRepositoryPolicies() ([]RepositoryPolicy, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return nil, err
-	}
-	var records []repositoryPolicyRecord
-	if err := db.Order("repository_full_name ASC, profile_name ASC, runner_group_name ASC").Find(&records).Error; err != nil {
-		return nil, err
-	}
-	policies := make([]RepositoryPolicy, 0, len(records))
-	for _, record := range records {
-		policies = append(policies, recordToRepositoryPolicy(record))
-	}
-	return policies, nil
-}
-
-func (s *DBStore) GetRepositoryPolicy(id int64) (RepositoryPolicy, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return RepositoryPolicy{}, err
-	}
-	var record repositoryPolicyRecord
-	if err := db.First(&record, "id = ?", id).Error; err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return RepositoryPolicy{}, ErrNotFound
-		}
-		return RepositoryPolicy{}, err
-	}
-	return recordToRepositoryPolicy(record), nil
-}
-
-func (s *DBStore) UpsertRepositoryPolicy(policy RepositoryPolicy) (RepositoryPolicy, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return RepositoryPolicy{}, err
-	}
-	policy.RepositoryFullName = strings.TrimSpace(policy.RepositoryFullName)
-	policy.ProfileName = strings.TrimSpace(policy.ProfileName)
-	policy.RunnerGroupName = strings.TrimSpace(policy.RunnerGroupName)
-	if policy.RepositoryFullName == "" {
-		return RepositoryPolicy{}, fmt.Errorf("repository_full_name is required")
-	}
-	if (policy.ProfileName == "") == (policy.RunnerGroupName == "") {
-		return RepositoryPolicy{}, fmt.Errorf("exactly one of runner_spec_name or runner_group_name is required")
-	}
-	now := time.Now().UTC()
-	if policy.ID == 0 {
-		record := repositoryPolicyRecord{
-			RepositoryFullName: policy.RepositoryFullName,
-			ProfileName:        policy.ProfileName,
-			RunnerGroupName:    policy.RunnerGroupName,
-			Enabled:            policy.Enabled,
-			CreatedAt:          now,
-		}
-		if record.ProfileName != "" {
-			var saved repositoryPolicyRecord
-			err := db.First(&saved, "repository_full_name = ? AND profile_name = ?", record.RepositoryFullName, record.ProfileName).Error
-			if err == nil {
-				saved.Enabled = record.Enabled
-				saved.RunnerGroupName = ""
-				if err := db.Save(&saved).Error; err != nil {
-					return RepositoryPolicy{}, err
-				}
-				return recordToRepositoryPolicy(saved), nil
-			}
-			if !errors.Is(err, gorm.ErrRecordNotFound) {
-				return RepositoryPolicy{}, err
-			}
-			if err := db.Create(&record).Error; err != nil {
-				return RepositoryPolicy{}, err
-			}
-			if err := db.First(&saved, "repository_full_name = ? AND profile_name = ?", record.RepositoryFullName, record.ProfileName).Error; err != nil {
-				return RepositoryPolicy{}, err
-			}
-			return recordToRepositoryPolicy(saved), nil
-		}
-		var saved repositoryPolicyRecord
-		err := db.First(&saved, "repository_full_name = ? AND runner_group_name = ?", record.RepositoryFullName, record.RunnerGroupName).Error
-		if err == nil {
-			saved.Enabled = record.Enabled
-			saved.ProfileName = ""
-			if err := db.Save(&saved).Error; err != nil {
-				return RepositoryPolicy{}, err
-			}
-			return recordToRepositoryPolicy(saved), nil
-		}
-		if !errors.Is(err, gorm.ErrRecordNotFound) {
-			return RepositoryPolicy{}, err
-		}
-		if err := db.Create(&record).Error; err != nil {
-			return RepositoryPolicy{}, err
-		}
-		if err := db.First(&saved, "repository_full_name = ? AND runner_group_name = ?", record.RepositoryFullName, record.RunnerGroupName).Error; err != nil {
-			return RepositoryPolicy{}, err
-		}
-		return recordToRepositoryPolicy(saved), nil
-	}
-	updates := map[string]any{
-		"repository_full_name": policy.RepositoryFullName,
-		"profile_name":         policy.ProfileName,
-		"runner_group_name":    policy.RunnerGroupName,
-		"enabled":              policy.Enabled,
-	}
-	if err := db.Model(&repositoryPolicyRecord{}).Where("id = ?", policy.ID).Updates(updates).Error; err != nil {
-		return RepositoryPolicy{}, err
-	}
-	var saved repositoryPolicyRecord
-	if err := db.First(&saved, "id = ?", policy.ID).Error; err != nil {
-		return RepositoryPolicy{}, err
-	}
-	return recordToRepositoryPolicy(saved), nil
-}
-
-func (s *DBStore) DeleteRepositoryPolicy(id int64) error {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return err
-	}
-	return db.Delete(&repositoryPolicyRecord{}, "id = ?", id).Error
-}
-
-func (s *DBStore) MatchProfile(repositoryFullName string, labels []string) (ProfileMatch, error) {
-	policies, err := s.ListRepositoryPolicies()
-	if err != nil {
-		return ProfileMatch{}, err
-	}
-	profiles, err := s.ListProfiles()
-	if err != nil {
-		return ProfileMatch{}, err
-	}
-	match := ProfileMatch{
-		RepositoryFullName: repositoryFullName,
-		Labels:             append([]string(nil), labels...),
-	}
-	allowed := map[string]bool{}
-	for _, profile := range profiles {
-		if profile.Enabled && profile.DefaultAvailable {
-			allowed[profile.Name] = true
-		}
-	}
-	for _, policy := range policies {
-		if !policy.Enabled {
-			continue
-		}
-		if repositoryMatches(policy.RepositoryFullName, repositoryFullName) {
-			if policy.ProfileName != "" {
-				allowed[policy.ProfileName] = true
-			}
-			if policy.RunnerGroupName != "" {
-				group, err := s.GetRunnerGroup(policy.RunnerGroupName)
-				if err != nil {
-					continue
-				}
-				if !group.Enabled {
-					continue
-				}
-				for _, specName := range group.SpecNames {
-					allowed[specName] = true
-				}
-			}
-		}
-	}
-	if len(allowed) == 0 {
-		match.Reason = "profile_not_allowed"
-		return match, nil
-	}
-	var candidates []RunnerProfile
-	for _, profile := range profiles {
-		if !profile.Enabled || !allowed[profile.Name] {
-			continue
-		}
-		if labelsMatch(labels, profile.Labels) {
-			candidates = append(candidates, profile)
-		}
-	}
-	if len(candidates) == 0 {
-		match.Reason = "profile_labels_not_matched"
-		return match, nil
-	}
-	sort.Slice(candidates, func(i, j int) bool {
-		if candidates[i].Priority != candidates[j].Priority {
-			return candidates[i].Priority > candidates[j].Priority
-		}
-		if len(candidates[i].Labels) != len(candidates[j].Labels) {
-			return len(candidates[i].Labels) > len(candidates[j].Labels)
-		}
-		return candidates[i].Name < candidates[j].Name
-	})
-	match.Profile = &candidates[0]
-	return match, nil
-}
-
-func (s *DBStore) GetAccountByOAuthIdentity(provider, subject string) (Account, OAuthIdentity, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return Account{}, OAuthIdentity{}, err
-	}
-	provider = normalizeOAuthProvider(provider)
-	subject = normalizeOAuthSubject(subject)
-	if provider == "" || subject == "" {
-		return Account{}, OAuthIdentity{}, ErrNotFound
-	}
-	var identity oauthIdentityRecord
-	if err := db.First(&identity, "oauth_provider = ? AND oauth_subject = ?", provider, subject).Error; err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return Account{}, OAuthIdentity{}, ErrNotFound
-		}
-		return Account{}, OAuthIdentity{}, err
-	}
-	return s.accountFromIdentity(db, identity)
-}
-
-func (s *DBStore) UpsertAccountForOAuthIdentity(identity OAuthIdentity, role string) (Account, OAuthIdentity, error) {
-	return s.saveOAuthIdentity(identity, role, true, 0)
-}
-
-func (s *DBStore) EnsureAccountForOAuthIdentity(identity OAuthIdentity, role string) (Account, OAuthIdentity, error) {
-	return s.saveOAuthIdentity(identity, role, false, 0)
-}
-
-func (s *DBStore) LinkOAuthIdentityToAccount(accountID int64, identity OAuthIdentity) (Account, OAuthIdentity, error) {
-	return s.saveOAuthIdentity(identity, "", false, accountID)
-}
-
-func (s *DBStore) saveOAuthIdentity(identity OAuthIdentity, role string, updateExisting bool, accountID int64) (Account, OAuthIdentity, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return Account{}, OAuthIdentity{}, err
-	}
-	provider := normalizeOAuthProvider(identity.OAuthProvider)
-	subject := normalizeOAuthSubject(identity.OAuthSubject)
-	login := normalizeOAuthLogin(identity.OAuthLogin)
-	if provider == "" {
-		return Account{}, OAuthIdentity{}, fmt.Errorf("oauth_provider is required")
-	}
-	if subject == "" {
-		return Account{}, OAuthIdentity{}, fmt.Errorf("oauth_subject is required")
-	}
-	if login == "" {
-		return Account{}, OAuthIdentity{}, fmt.Errorf("oauth_login is required")
-	}
-	role = normalizePlatformRole(role)
-	if accountID == 0 && role == "" {
-		return Account{}, OAuthIdentity{}, fmt.Errorf("role must be admin or user")
-	}
-	var savedAccount Account
-	var savedIdentity OAuthIdentity
-	var lastErr error
-	for attempt := 0; attempt < 20; attempt++ {
-		savedAccount, savedIdentity, lastErr = s.saveOAuthIdentityOnce(db, provider, subject, login, role, identity.CreatedAt, updateExisting, accountID)
-		if lastErr == nil {
-			return savedAccount, savedIdentity, nil
-		}
-		if !isTransientStoreError(lastErr) {
-			return Account{}, OAuthIdentity{}, lastErr
-		}
-		time.Sleep(time.Duration(attempt+1) * 25 * time.Millisecond)
-	}
-	return Account{}, OAuthIdentity{}, lastErr
-}
-
-func (s *DBStore) saveOAuthIdentityOnce(db *gorm.DB, provider, subject, login, role string, createdAt time.Time, updateExisting bool, accountID int64) (Account, OAuthIdentity, error) {
-	now := time.Now().UTC()
-	var savedAccount Account
-	var savedIdentity OAuthIdentity
-	err := db.Transaction(func(tx *gorm.DB) error {
-		var identity oauthIdentityRecord
-		err := tx.First(&identity, "oauth_provider = ? AND oauth_subject = ?", provider, subject).Error
-		if err == nil {
-			if accountID != 0 && identity.AccountID != accountID {
-				return ErrConflict
-			}
-			updates := map[string]any{
-				"oauth_login": login,
-				"updated_at":  now,
-			}
-			if err := tx.Model(&identity).Updates(updates).Error; err != nil {
-				return err
-			}
-			if updateExisting {
-				if err := tx.Model(&accountRecord{}).Where("id = ?", identity.AccountID).Updates(map[string]any{
-					"role":       role,
-					"updated_at": now,
-				}).Error; err != nil {
-					return err
-				}
-			}
-			identity.OAuthLogin = login
-			identity.UpdatedAt = now
-			var identityErr error
-			savedAccount, savedIdentity, identityErr = s.accountFromIdentity(tx, identity)
-			return identityErr
-		}
-		if !errors.Is(err, gorm.ErrRecordNotFound) {
-			return err
-		}
-
-		targetAccountID := accountID
-		createdAccountID := int64(0)
-		if targetAccountID == 0 {
-			account := accountRecord{
-				Role:      role,
-				CreatedAt: now,
-				UpdatedAt: now,
-			}
-			if err := tx.Create(&account).Error; err != nil {
-				return err
-			}
-			targetAccountID = account.ID
-			createdAccountID = account.ID
-		} else {
-			var account accountRecord
-			if err := tx.First(&account, "id = ?", targetAccountID).Error; err != nil {
-				if errors.Is(err, gorm.ErrRecordNotFound) {
-					return ErrNotFound
-				}
-				return err
-			}
-		}
-		identity = oauthIdentityRecord{
-			AccountID:     targetAccountID,
-			OAuthProvider: provider,
-			OAuthSubject:  subject,
-			OAuthLogin:    login,
-			CreatedAt:     now,
-			UpdatedAt:     now,
-		}
-		if !createdAt.IsZero() {
-			identity.CreatedAt = createdAt
-		}
-		result := tx.Clauses(clause.OnConflict{DoNothing: true}).Create(&identity)
-		if result.Error != nil {
-			return result.Error
-		}
-		if result.RowsAffected == 0 {
-			if createdAccountID != 0 {
-				if err := tx.Delete(&accountRecord{}, "id = ?", createdAccountID).Error; err != nil {
-					return err
-				}
-			}
-			if err := tx.First(&identity, "oauth_provider = ? AND oauth_subject = ?", provider, subject).Error; err != nil {
-				return err
-			}
-			if accountID != 0 && identity.AccountID != accountID {
-				return ErrConflict
-			}
-			if err := tx.Model(&identity).Updates(map[string]any{
-				"oauth_login": login,
-				"updated_at":  now,
-			}).Error; err != nil {
-				return err
-			}
-			if updateExisting {
-				if err := tx.Model(&accountRecord{}).Where("id = ?", identity.AccountID).Updates(map[string]any{
-					"role":       role,
-					"updated_at": now,
-				}).Error; err != nil {
-					return err
-				}
-			}
-			identity.OAuthLogin = login
-			identity.UpdatedAt = now
-		}
-		var identityErr error
-		savedAccount, savedIdentity, identityErr = s.accountFromIdentity(tx, identity)
-		return identityErr
-	})
-	if err != nil {
-		return Account{}, OAuthIdentity{}, err
-	}
-	return savedAccount, savedIdentity, nil
-}
-
-func isTransientStoreError(err error) bool {
-	if err == nil {
-		return false
-	}
-	message := strings.ToLower(err.Error())
-	return strings.Contains(message, "database is locked") ||
-		strings.Contains(message, "sqlite_busy") ||
-		strings.Contains(message, "sqlstate 40001") ||
-		strings.Contains(message, "sqlstate 40p01") ||
-		strings.Contains(message, "serialization_failure") ||
-		strings.Contains(message, "deadlock_detected") ||
-		strings.Contains(message, "concurrent update")
-}
-
-func (s *DBStore) AppendAuditEvent(event AuditEvent) (AuditEvent, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return AuditEvent{}, err
-	}
-	record := auditEventRecord{
-		Actor:        strings.TrimSpace(event.Actor),
-		Action:       strings.TrimSpace(event.Action),
-		ResourceType: strings.TrimSpace(event.ResourceType),
-		ResourceID:   strings.TrimSpace(event.ResourceID),
-		PayloadJSON:  event.PayloadJSON,
-		CreatedAt:    event.CreatedAt.UTC(),
-	}
-	if record.CreatedAt.IsZero() {
-		record.CreatedAt = time.Now().UTC()
-	}
-	if err := db.Create(&record).Error; err != nil {
-		return AuditEvent{}, err
-	}
-	return AuditEvent(record), nil
-}
-
-func (s *DBStore) ListAuditEvents(limit int) ([]AuditEvent, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return nil, err
-	}
-	if limit <= 0 {
-		limit = 100
-	}
-	var records []auditEventRecord
-	if err := db.Order("created_at DESC").Limit(limit).Find(&records).Error; err != nil {
-		return nil, err
-	}
-	events := make([]AuditEvent, 0, len(records))
-	for _, record := range records {
-		events = append(events, AuditEvent(record))
-	}
-	return events, nil
-}
-
-func (s *DBStore) dbOrEnsure() (*gorm.DB, error) {
-	if err := s.Ensure(); err != nil {
-		return nil, err
-	}
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.db, nil
-}
-
-func (s *DBStore) readRecord(id string) (runnerRequestRecord, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return runnerRequestRecord{}, err
-	}
-	var record runnerRequestRecord
-	if err := db.First(&record, "id = ?", sanitizeID(id)).Error; err != nil {
-		return runnerRequestRecord{}, err
-	}
-	return record, nil
-}
-
-func (s *DBStore) countStatesForProfile(name string, statuses []string) (int, error) {
-	db, err := s.dbOrEnsure()
-	if err != nil {
-		return 0, err
-	}
-	var count int64
-	if err := db.Model(&runnerRequestRecord{}).
-		Where("profile_name = ?", strings.TrimSpace(name)).
-		Where("status IN ?", statuses).
-		Count(&count).Error; err != nil {
-		return 0, err
-	}
-	return int(count), nil
-}
-
-func (s *DBStore) open() (*gorm.DB, error) {
-	cfg := &gorm.Config{Logger: gormlogger.Default.LogMode(gormlogger.Silent)}
-	switch s.opts.Backend {
-	case BackendSQLite:
-		dir := filepath.Dir(s.opts.DatabaseURL)
-		if dir != "." && dir != "" {
-			if err := os.MkdirAll(dir, 0o755); err != nil {
-				return nil, err
-			}
-		}
-		db, err := gorm.Open(sqlite.Open(s.opts.DatabaseURL), cfg)
-		if err != nil {
-			return nil, err
-		}
-		if err := configureSQLite(db); err != nil {
-			return nil, err
-		}
-		return db, nil
-	case BackendPostgres:
-		return gorm.Open(postgres.Open(s.opts.DatabaseURL), cfg)
-	default:
-		return nil, fmt.Errorf("unsupported state backend: %s", s.opts.Backend)
-	}
-}
-
-func configureSQLite(db *gorm.DB) error {
-	sqlDB, err := db.DB()
-	if err != nil {
-		return err
-	}
-	sqlDB.SetMaxOpenConns(1)
-	sqlDB.SetMaxIdleConns(1)
-	pragmas := []string{
-		"PRAGMA foreign_keys = ON",
-		"PRAGMA journal_mode = WAL",
-		"PRAGMA synchronous = NORMAL",
-		"PRAGMA busy_timeout = 15000",
-	}
-	for _, pragma := range pragmas {
-		if err := db.Exec(pragma).Error; err != nil {
-			return fmt.Errorf("%s: %w", pragma, err)
-		}
-	}
-	return nil
-}
-
-func (s *DBStore) migrate(db *gorm.DB) error {
-	if err := migrateLegacySchemaColumns(db); err != nil {
-		return err
-	}
-	return db.AutoMigrate(
-		&runnerRequestRecord{},
-		&runnerEventRecord{},
-		&runnerProfileRecord{},
-		&runnerGroupRecord{},
-		&runnerGroupSpecRecord{},
-		&repositoryPolicyRecord{},
-		&auditEventRecord{},
-		&accountRecord{},
-		&oauthIdentityRecord{},
-	)
-}
-
-func migrateLegacySchemaColumns(db *gorm.DB) error {
-	migrator := db.Migrator()
-	if migrator.HasTable(&runnerProfileRecord{}) && !migrator.HasColumn(&runnerProfileRecord{}, "default_available") {
-		if err := migrator.AddColumn(&legacyRunnerProfileDefaultAvailableColumn{}, "DefaultAvailable"); err != nil {
-			return err
-		}
-	}
-	if migrator.HasTable(&repositoryPolicyRecord{}) && !migrator.HasColumn(&repositoryPolicyRecord{}, "runner_group_name") {
-		if err := migrator.AddColumn(&legacyRepositoryPolicyRunnerGroupNameColumn{}, "RunnerGroupName"); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func applyStateTimestamps(st *RunnerState, now time.Time) {
-	switch st.Status {
-	case StatusCreating:
-		if st.CreatingAt.IsZero() {
-			st.CreatingAt = now
-		}
-	case StatusRunning:
-		if st.CreatingAt.IsZero() {
-			st.CreatingAt = now
-		}
-		if st.RunningAt.IsZero() {
-			st.RunningAt = now
-		}
-	case StatusStopping:
-		if st.StoppingAt.IsZero() {
-			st.StoppingAt = now
-		}
-	case StatusCompleted:
-		if st.StoppingAt.IsZero() {
-			st.StoppingAt = now
-		}
-		if st.CompletedAt.IsZero() {
-			st.CompletedAt = now
-		}
-	case StatusFailed:
-		if st.FailedAt.IsZero() {
-			st.FailedAt = now
-		}
-	}
-}
-
-func recordToRequest(record runnerRequestRecord) (RunnerRequest, error) {
-	var labels []string
-	if err := json.Unmarshal([]byte(record.LabelsJSON), &labels); err != nil {
-		return RunnerRequest{}, err
-	}
-	var requestedLabels []string
-	if record.RequestedLabelsJSON != "" {
-		if err := json.Unmarshal([]byte(record.RequestedLabelsJSON), &requestedLabels); err != nil {
-			return RunnerRequest{}, err
-		}
-	}
-	return RunnerRequest{
-		ID:                 record.ID,
-		Source:             record.Source,
-		JobID:              pointerToInt64(record.WorkflowJobID),
-		RepositoryFullName: record.RepositoryFullName,
-		RequestedLabels:    requestedLabels,
-		Labels:             labels,
-		ProfileName:        record.ProfileName,
-		RunnerGroup:        record.RunnerGroup,
-		RunnerName:         record.RunnerName,
-		CreatedAt:          record.QueuedAt,
-	}, nil
-}
-
-func recordToState(record runnerRequestRecord) RunnerState {
-	requestedLabels, _ := labelsFromJSON(record.RequestedLabelsJSON)
-	githubLinks := githubLinksFromPayload(record)
-	return RunnerState{
-		ID:                 record.ID,
-		Status:             record.Status,
-		RepositoryFullName: record.RepositoryFullName,
-		RequestedLabels:    requestedLabels,
-		ProfileName:        record.ProfileName,
-		RunnerGroup:        record.RunnerGroup,
-		RunnerName:         record.RunnerName,
-		SandboxID:          record.SandboxID,
-		ProcessPID:         record.ProcessPID,
-		WorkflowJobID:      pointerToInt64(record.WorkflowJobID),
-		WorkflowRunID:      githubLinks.workflowRunID,
-		GitHubJobURL:       githubLinks.jobURL,
-		PullRequestNumber:  githubLinks.pullRequestNumber,
-		AssignedJobID:      record.AssignedJobID,
-		AssignedJobName:    record.AssignedJobName,
-		Error:              record.Error,
-		FailureStage:       record.FailureStage,
-		FailureReason:      record.FailureReason,
-		LastErrorCode:      record.LastErrorCode,
-		LastErrorMessage:   record.LastErrorMessage,
-		LastErrorRetryable: record.LastErrorRetryable,
-		RetryCount:         record.RetryCount,
-		UpdatedAt:          record.UpdatedAt,
-		CreatedAt:          record.QueuedAt,
-		LastAttemptAt:      pointerToTime(record.LastAttemptAt),
-		NextRetryAt:        pointerToTime(record.NextRetryAt),
-		CreatingAt:         pointerToTime(record.CreatingAt),
-		RunningAt:          pointerToTime(record.RunningAt),
-		StoppingAt:         pointerToTime(record.StoppingAt),
-		CompletedAt:        pointerToTime(record.CompletedAt),
-		FailedAt:           pointerToTime(record.FailedAt),
-		LeaseOwner:         record.LeaseOwner,
-		LeaseExpiresAt:     pointerToTime(record.LeaseExpiresAt),
-		Version:            record.Version,
-	}
-}
-
-func (s *DBStore) accountFromIdentity(db *gorm.DB, identity oauthIdentityRecord) (Account, OAuthIdentity, error) {
-	var account accountRecord
-	if err := db.First(&account, "id = ?", identity.AccountID).Error; err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return Account{}, OAuthIdentity{}, ErrNotFound
-		}
-		return Account{}, OAuthIdentity{}, err
-	}
-	return Account(account), recordToOAuthIdentity(identity), nil
-}
-
-func recordToOAuthIdentity(record oauthIdentityRecord) OAuthIdentity {
-	return OAuthIdentity{
-		ID:            record.ID,
-		AccountID:     record.AccountID,
-		OAuthProvider: record.OAuthProvider,
-		OAuthSubject:  record.OAuthSubject,
-		OAuthLogin:    record.OAuthLogin,
-		CreatedAt:     record.CreatedAt,
-		UpdatedAt:     record.UpdatedAt,
-	}
-}
-
-type githubPayloadLinks struct {
-	workflowRunID     int64
-	jobURL            string
-	pullRequestNumber int64
-}
-
-type githubPayloadPullRequest struct {
-	Number int64 `json:"number"`
-}
-
-func githubLinksFromPayload(record runnerRequestRecord) githubPayloadLinks {
-	var payload struct {
-		WorkflowJob struct {
-			RunID        int64                      `json:"run_id"`
-			HTMLURL      string                     `json:"html_url"`
-			PullRequests []githubPayloadPullRequest `json:"pull_requests"`
-			WorkflowRun  struct {
-				ID int64 `json:"id"`
-			} `json:"workflow_run"`
-		} `json:"workflow_job"`
-		WorkflowRun struct {
-			ID           int64                      `json:"id"`
-			HTMLURL      string                     `json:"html_url"`
-			PullRequests []githubPayloadPullRequest `json:"pull_requests"`
-		} `json:"workflow_run"`
-		PullRequest struct {
-			Number int64 `json:"number"`
-		} `json:"pull_request"`
-	}
-	if record.GitHubPayloadJSON != "" {
-		_ = json.Unmarshal([]byte(record.GitHubPayloadJSON), &payload)
-	}
-
-	runID := payload.WorkflowJob.RunID
-	if runID == 0 {
-		runID = payload.WorkflowJob.WorkflowRun.ID
-	}
-	if runID == 0 {
-		runID = payload.WorkflowRun.ID
-	}
-
-	prNumber := firstPullRequestNumber(payload.WorkflowJob.PullRequests)
-	if prNumber == 0 {
-		prNumber = firstPullRequestNumber(payload.WorkflowRun.PullRequests)
-	}
-	if prNumber == 0 {
-		prNumber = payload.PullRequest.Number
-	}
-
-	jobURL := payload.WorkflowJob.HTMLURL
-	jobID := record.AssignedJobID
-	if jobID == 0 {
-		jobID = pointerToInt64(record.WorkflowJobID)
-	}
-	if jobURL == "" && record.RepositoryFullName != "" && runID > 0 && jobID > 0 {
-		jobURL = fmt.Sprintf("https://github.com/%s/actions/runs/%d/job/%d", record.RepositoryFullName, runID, jobID)
-	}
-	jobURL = appendPullRequestQuery(jobURL, prNumber)
-
-	return githubPayloadLinks{
-		workflowRunID:     runID,
-		jobURL:            jobURL,
-		pullRequestNumber: prNumber,
-	}
-}
-
-func firstPullRequestNumber(pulls []githubPayloadPullRequest) int64 {
-	if len(pulls) == 0 {
-		return 0
-	}
-	return pulls[0].Number
-}
-
-func appendPullRequestQuery(rawURL string, prNumber int64) string {
-	if rawURL == "" || prNumber == 0 || strings.Contains(rawURL, "pr=") {
-		return rawURL
-	}
-	separator := "?"
-	if strings.Contains(rawURL, "?") {
-		separator = "&"
-	}
-	return fmt.Sprintf("%s%spr=%d", rawURL, separator, prNumber)
-}
-
-func recordToProfile(record runnerProfileRecord) (RunnerProfile, error) {
-	var labels []string
-	if err := json.Unmarshal([]byte(record.LabelsJSON), &labels); err != nil {
-		return RunnerProfile{}, err
-	}
-	return RunnerProfile{
-		Name:             record.Name,
-		Labels:           labels,
-		TemplateID:       record.TemplateID,
-		RunnerGroup:      record.RunnerGroup,
-		MaxConcurrency:   record.MaxConcurrency,
-		MinIdle:          record.MinIdle,
-		Priority:         record.Priority,
-		Enabled:          record.Enabled,
-		DefaultAvailable: record.DefaultAvailable,
-		CreatedAt:        record.CreatedAt,
-		UpdatedAt:        record.UpdatedAt,
-	}, nil
-}
-
-func (s *DBStore) recordToRunnerGroup(db *gorm.DB, record runnerGroupRecord) (RunnerGroup, error) {
-	var links []runnerGroupSpecRecord
-	if err := db.Where("group_name = ?", record.Name).Order("spec_name ASC").Find(&links).Error; err != nil {
-		return RunnerGroup{}, err
-	}
-	specNames := make([]string, 0, len(links))
-	for _, link := range links {
-		specNames = append(specNames, link.SpecName)
-	}
-	return RunnerGroup{
-		Name:        record.Name,
-		Description: record.Description,
-		SpecNames:   specNames,
-		Enabled:     record.Enabled,
-		CreatedAt:   record.CreatedAt,
-		UpdatedAt:   record.UpdatedAt,
-	}, nil
-}
-
-func recordToRepositoryPolicy(record repositoryPolicyRecord) RepositoryPolicy {
-	return RepositoryPolicy(record)
-}
-
-func uniqueTrimmed(values []string) []string {
-	seen := map[string]bool{}
-	out := make([]string, 0, len(values))
-	for _, value := range values {
-		value = strings.TrimSpace(value)
-		if value == "" || seen[value] {
-			continue
-		}
-		seen[value] = true
-		out = append(out, value)
-	}
-	return out
-}
-
-func zeroTimeToPointer(t time.Time) *time.Time {
-	if t.IsZero() {
-		return nil
-	}
-	tt := t.UTC()
-	return &tt
-}
-
-func pointerToTime(t *time.Time) time.Time {
-	if t == nil {
-		return time.Time{}
-	}
-	return t.UTC()
-}
-
-func pointerToInt64(v *int64) int64 {
-	if v == nil {
-		return 0
-	}
-	return *v
-}
-
-func labelsFromJSON(data string) ([]string, error) {
-	if strings.TrimSpace(data) == "" {
-		return nil, nil
-	}
-	var labels []string
-	if err := json.Unmarshal([]byte(data), &labels); err != nil {
-		return nil, err
-	}
-	return labels, nil
-}
-
-func repositoryMatches(pattern, repository string) bool {
-	pattern = strings.TrimSpace(pattern)
-	repository = strings.TrimSpace(repository)
-	if pattern == "" || repository == "" {
-		return false
-	}
-	if pattern == repository {
-		return true
-	}
-	if strings.HasSuffix(pattern, "/*") {
-		return strings.HasPrefix(repository, strings.TrimSuffix(pattern, "*"))
-	}
-	return false
-}
-
-func labelsMatch(jobLabels, required []string) bool {
-	return labelutil.Match(jobLabels, required)
-}
-
-func logEventType(name string) (string, error) {
-	switch filepath.Base(name) {
-	case "control.log":
-		return "control_log", nil
-	case "stdout.log":
-		return "stdout_log", nil
-	case "stderr.log":
-		return "stderr_log", nil
-	default:
-		return "", fmt.Errorf("unsupported log name")
-	}
-}
-
-func sanitizeID(id string) string {
-	id = strings.TrimSpace(id)
-	id = strings.ReplaceAll(id, "/", "-")
-	id = strings.ReplaceAll(id, "\\", "-")
-	id = strings.ReplaceAll(id, "..", "-")
-	return id
-}
-
-func sanitizeRunnerName(name string) string {
-	name = strings.TrimSpace(name)
-	name = strings.ReplaceAll(name, "/", "-")
-	name = strings.ReplaceAll(name, "\\", "-")
-	return name
-}
-
-func normalizeOAuthProvider(provider string) string {
-	return strings.ToLower(strings.TrimSpace(provider))
-}
-
-func normalizeOAuthSubject(subject string) string {
-	return strings.TrimSpace(subject)
-}
-
-func normalizeOAuthLogin(login string) string {
-	return strings.ToLower(strings.TrimSpace(login))
-}
-
-func normalizePlatformRole(role string) string {
-	role = strings.ToLower(strings.TrimSpace(role))
-	switch role {
-	case "admin", "user":
-		return role
-	default:
-		return ""
-	}
 }
